GIBS-1008 Use mp_safe only for MRFs with z-levels

+++ src/mrfgen/mrfgen.py
@@ -1383,16 +1383,17 @@ else:
     # Send to log.
     log_info_mssg(str().join(['size of MRF:  ', sizeX, ' x ', sizeY]))
     
-    # Add mp_safe to Raster
-    mrf_file.seek(0)
-    lines = mrf_file.readlines()
-    for idx in range(0, len(lines)):
-        if '<Raster' in lines[idx]:
-            lines[idx] = lines[idx].replace('<Raster','<Raster mp_safe="on"')
-            log_info_mssg("Set MRF mp_safe on")
-    mrf_file.seek(0)
-    mrf_file.truncate()
-    mrf_file.writelines(lines)
+    # Add mp_safe to Raster if using z levels
+    if zlevels != '' and zkey != '':
+        mrf_file.seek(0)
+        lines = mrf_file.readlines()
+        for idx in range(0, len(lines)):
+            if '<Raster' in lines[idx]:
+                lines[idx] = lines[idx].replace('<Raster','<Raster mp_safe="on"')
+                log_info_mssg("Set MRF mp_safe on")
+        mrf_file.seek(0)
+        mrf_file.truncate()
+        mrf_file.writelines(lines)
     
     # Close file.
     mrf_file.close()
